Add new checker: Or(ch1, ch2, ...)
This adds a new checker, Or, which allows multiple checkers to be
combined using a boolean "or" instead of the default "and". Validation
is successful if one of the supplied checkers is valid. Checkers are
executed in order from left to right, and the first success returns
a valid response. Therefore, it's a good idea to put expensive checkers
on the right.

This may be useful for composing generic checkers to acheive a result
where writing a custom checker would have been required previously.

Example:

// phone number is either empty or must be valid if specified
vala.BeginValidation().Validate(
    vala.Or(
	vala.HasLen(phoneNum, 0, "phoneNumber"),
	isValidPhoneNumber(phoneNum, "phoneNumber")
    )
)